rename 'test/olm' to 'test/assert'
