+++ pkg/controller/cheinstallation/cheinstallation_controller_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/codeready-toolchain/toolchain-operator/pkg/apis/toolchain/v1alpha1"
 	"github.com/codeready-toolchain/toolchain-operator/pkg/toolchain"
 	"github.com/codeready-toolchain/toolchain-operator/test"
-	. "github.com/codeready-toolchain/toolchain-operator/test/olm"
+	. "github.com/codeready-toolchain/toolchain-operator/test/assert"
 
 	orgv1 "github.com/eclipse/che-operator/pkg/apis/org/v1"
 	"github.com/go-logr/logr"
