+++ test/assert/che_cluster.go
@@ -1,4 +1,4 @@
-package olm
+package assert
 
 import (
 	"context"
