+++ test/assert/che_installation.go
@@ -1,4 +1,4 @@
-package olm
+package assert
 
 import (
 	"context"
