+++ test/assert/subscription.go
@@ -1,4 +1,4 @@
-package olm
+package assert
 
 import (
 	"context"
